==> Clasificación de spam/notebooks/01_exploracion_analisis_limpieza.ipynb <==
#Cargamos los datos para entener los datos
import pandas as pd 

#Usamos el enconding porque problablemente el dataset viene con caracteres que en UTF-8 no carga , evitamos error 
df = pd.read_csv('../data/spam.csv', encoding='latin-1', sep=';', names=['clase', 'mensajes'])
df.head(2)

(HAM) = No Spam , (SPAM) = Spam

#Miramos cantidad de ejemplos que tenemos
df.shape

#Revisamos que los datos sean los correctos
df.dtypes

#Revisamos nulos
df.isnull().sum()

#Revisamos balanceo en la variable clase
df['clase'].value_counts(normalize=True)

#Revisamos duplicados 
df.duplicated().sum()

#Revisamos valores duplicados
df[df.duplicated()].head(10)

#Eliminamos los duplicados para evitar multicolonealidad
df.drop_duplicates(inplace=True)

#Rectificamos que ya no tengamos duplicados
df.duplicated().sum()

#Revisamos longitud de los para poder revisar cual de los textos es mas largo si los spam o los ham
df['longitud'] = df['mensajes'].apply(len)
df.groupby('clase')['longitud'].mean()

Podemos ver como en promedio los que tienen spam son textos más largos 

#Creamos un nuevo dataset , sin los datos duplicaods, con nuevas columnas lito para hacer los procedimientos de entrnamiento
df.to_csv('../data/spam_clean.csv', index=False)

### Conclusiones del análisis exploratorio

- Revisamos la **cantidad de ejemplos** disponibles para confirmar que contamos con una base de datos suficiente.  
- Verificamos que los datos estuvieran en el **tipo de dato correspondiente** y que no existieran valores nulos.  
- Analizamos el **balanceo de clases**, encontrando una distribución desigual con **13.4% de Spam** y **86.6% de Ham**.  
  - Esto nos indica un desbalance que probablemente tengamos que tratar más adelante, dependiendo de cómo se comporten las métricas (como **F1-score**) durante el entrenamiento.  
  - En ese caso, podríamos aplicar técnicas de balanceo como **oversampling** o **undersampling**.  
- Tratamos los **valores duplicados** para evitar multicolinealidad.  
- Revisamos la **longitud de los mensajes**, observando que en muchos casos los mensajes de spam tienden a ser más largos que los de ham.
