==> Clasificación de spam/notebooks/02_test_train_validation.ipynb <==
#Imporatmos librerías necesarias
import pandas as pd
from sklearn.model_selection import train_test_split #Separar los datos
from sklearn.linear_model import LogisticRegression #Modelo de clasificacion binaria
from sklearn.metrics  import confusion_matrix, classification_report, accuracy_score #Evluacion del modelo
from sklearn.feature_extraction.text import TfidfVectorizer #Vectorizar los textos

#Cargamos los datos
df = pd.read_csv('../data/spam_clean.csv')
df.head(2)

#Separamos los datos a utilizar y el dato a predecir 
X = df['mensajes']
y = df['clase']

#Dividimos nuestros datos en 80% Train y 20% Test para que no memorice
X_train, X_test, Y_train, Y_test = train_test_split(X, y, random_state=42, test_size=0.2, stratify=y) #Stratify para que tenga la misma cantidad de ejemplos en ambos lados 

#Codifico la variable objetivo poniendo 0 si no es spam y 1 si si
Y_train = Y_train.map({'ham': 0, 'spam': 1})
Y_test = Y_test.map({'ham': 0, 'spam': 1})

Explicación rapida de que hace el TF-IDF , que vamos a utilizar, este basicmanete se encarga de agarrar estos mensajes , separarlos por palabras y cada palabra va ser una posicion dentro del vector , luego se transforma cada una de esas palabras a numeros y ese numero va representar la importancia de esa palabra en el mensaje 

tfidf = TfidfVectorizer(
    lowercase=True, #Pasa todo a mayusculas
    strip_accents='unicode', #Algunos caracteres especiales los normaliza como la ñ
    stop_words='english', #Acá ignora palabras muy comunes del idioma 
    ngram_range=(1, 2), #Captura frases especiales como 'free entry'
    norm='l2', #Normaliza los vectores (Puede mejorar modelos lineales)
)

#Convierto a numeros los mensajes dentro de vectores
X_train = tfidf.fit_transform(X_train)
X_test = tfidf.transform(X_test)

#Llamamos el modelo de regresion logisitca
model = LogisticRegression()
model.fit(X_train, Y_train) #Ajusta la probabilidad predicha a la clase real lo mejor posible

#Le pasamos nuevos valores para que aaplique lo aprendido 
y_predict = model.predict(X_test)

print(f'Accuracy {accuracy_score(Y_test, y_predict)}')
print(f'Reporte general \n {classification_report(Y_test, y_predict)}')

#Vemos de los datos reales cuantos acierta y cuantos falla
print(confusion_matrix(Y_test, y_predict)) #De no spam acierta 902 y falla 2 , de spam acierta 82 y falla 48

En estas metricas podemos obeservar como aparentemente tenemos una buena precision general con un 95% de aciertos y como el modelo de 130 ejemplos acerto que era spam el 63% de ellos , sin embargo como podemos ver en la clase de no spam, acierta que no hay spam casi 100% lo cual se debe al desbalanceo tan grande que hay en datos entonces vamos a proceder con un oversampler para balancear un subiendo ejemplos de clase 1 a el max de clase 0

#Libreria de resample
#Utilizaremos smote , ya que este creea nuevos valores algo mas variables para 
from imblearn.over_sampling import SMOTE

#Creamos la instancia
smote = SMOTE(random_state=42)

#Creamos nuestras nuevas variables con los resamples
X_over, Y_over = smote.fit_resample(X_train, Y_train)

#Reentrnamos el algoritmo con los nuevos datos
model = LogisticRegression()
model.fit(X_over, Y_over)

#realizamos nuevamente la prediccion pero está vez con el modelo entrenado con nuevos datos para poder revisar como quedó
y_predict = model.predict(X_test)
print(f'Accuracy {accuracy_score(Y_test, y_predict)}')
print(f'Reporte general \n {classification_report(Y_test, y_predict)}')
print(f'Confussion matrix \n {confusion_matrix(Y_test, y_predict)}')

### En esta parte pudimos ver como entrenamos el modelo , procesando los mensajes conviertiendolos en vectores con indices de importancia y frecuencia para que la computadora pueda entenderlos, también vemos como el modelo se ve afectado por datos desbalanceados y como rebalanceamos el modelo para que este pudiera aprender mas ejemplos de spam y pudiera clasifciar mejor, obteniendo mejores metricas con valores neuevos y acertando mucho mejor cuando es spam reduciendo un poco cuando no lo es
