==> Clasificación de spam/notebooks/03_pipline_final.ipynb <==
# Crearemos ya nuestro modelo final para guardarlo y luego crear una parte interactiva

#Librerías bases 

import pandas as pd #Trabajar con dataframe
import joblib #Guarda trabajos pesados para luego cargarlos facilmente
from sklearn.model_selection import train_test_split #Separa los datos
from sklearn.linear_model import LogisticRegression #Modelo de clasificacion
from sklearn.feature_extraction.text import TfidfVectorizer #Convierte los mensajes a vectores numericos
from imblearn.over_sampling import SMOTE #Hace muestras artificiales con un poco de variacion para rellenar el desbalanceo de los datos
from imblearn.pipeline import Pipeline as ImbPipline #Me permite integrar el vectorizer el resmapleo y el modelo en uno
from sklearn.metrics import accuracy_score, classification_report #Metricas de evaluación 

#Cargamos nuestros datos 
df = pd.read_csv('../data/spam_clean.csv')
df.head(2)

#Preparamos las variables a utilizar
X = df['mensajes']
y = df['clase'] #Objetivo

#Separamos nuestros datos para permitir una generalización. 80% Train, 20% Test
X_train, X_test, Y_train, Y_test = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)

#Transformarmos nuestra clase
Y_train = Y_train.map({'ham': 0, 'spam': 1})
Y_test = Y_test.map({'ham': 0, 'spam':1})

#Creamos el pipline para poder manejar varias cosas al tiempo
pipline = ImbPipline([
    ('tfidf', TfidfVectorizer(lowercase=True,strip_accents='unicode',stop_words='english',ngram_range=(1, 2),norm='l2')),
    ('smote', SMOTE(random_state=42)),
    ('model', LogisticRegression())
])

pipline.fit(X_train, Y_train)

#Reevaluamos el modelo
y_predict = pipline.predict(X_test)

#Metricas
print(f'Accuracy {accuracy_score(Y_test, y_predict)}')
print(f'Reporte general {classification_report(Y_test, y_predict)}')

#Guardamos el modelo

joblib.dump(pipline, 'spam_detect.pkl')
print('Modelo guardado')

### En este notebook , implementamos un pipline para encadenar los diferntes metodos y que sea mucho mas clean al momento de procesaro todo, y guardamos nuestro modelo para cargarlo en un script para poder generar una interactividad
